+++ kit/react-select/index.js
@@ -6,36 +6,22 @@
  */
 import {elemFactory} from '@xh/hoist/core';
 import Select from 'react-select';
-import CreatableSelect from 'react-select/creatable';
 import AsyncSelect from 'react-select/async';
 import AsyncCreatable from 'react-select/async-creatable';
-
-import {makeAsyncSelect} from 'react-select/async';
-import {makeCreatableSelect} from 'react-select/creatable';
+import Creatable from 'react-select/creatable';
 import WindowedSelect from 'react-windowed-select';
-
-// Windowed support
-const WindowedAsyncSelect = makeAsyncSelect(WindowedSelect);
-const WindowedCreatableSelect = makeCreatableSelect(WindowedSelect);
-const WindowedAsyncCreatable = makeAsyncSelect(WindowedCreatableSelect);
 
 export {
     Select,
     AsyncSelect,
     AsyncCreatable,
-    CreatableSelect,
-    WindowedSelect,
-    WindowedAsyncSelect,
-    WindowedCreatableSelect,
-    WindowedAsyncCreatable
+    Creatable,
+    WindowedSelect
 };
 
 export const
     reactSelect = elemFactory(Select),
-    reactCreatableSelect = elemFactory(CreatableSelect),
+    reactCreatableSelect = elemFactory(Creatable),
     reactAsyncSelect = elemFactory(AsyncSelect),
     reactAsyncCreatableSelect = elemFactory(AsyncCreatable),
-    reactWindowedSelect = elemFactory(WindowedSelect),
-    reactWindowedCreatableSelect = elemFactory(WindowedCreatableSelect),
-    reactWindowedAsyncSelect = elemFactory(WindowedAsyncSelect),
-    reactWindowedAsyncCreatableSelect = elemFactory(WindowedAsyncCreatable);
+    reactWindowedSelect = elemFactory(WindowedSelect);